Prevent same library being loaded multiple times

+++ diozero-core/src/main/java/com/diozero/util/LibraryLoader.java
@@ -31,24 +31,33 @@ import java.io.IOException;
 import java.nio.file.Files;
 import java.nio.file.Path;
 import java.nio.file.StandardCopyOption;
+import java.util.HashMap;
+import java.util.Map;
 
 import org.pmw.tinylog.Logger;
 
 public class LibraryLoader {
+	private static final Object LOADED = new Object();
+	private static final Map<String, Object> LOADED_LIBRARIES = new HashMap<>();
+	
 	public static void loadLibrary(Class<?> clz, String libName) throws UnsatisfiedLinkError {
-		;
-		try {
-			// First try load the library from within the JAR file
-			Path path = Files.createTempFile("lib" + libName, ".so");
-			path.toFile().deleteOnExit();
-			Files.copy(clz.getResourceAsStream("/lib/" + SystemInfo.getLibraryPath() + "/lib" + libName + ".so"), path,
-					StandardCopyOption.REPLACE_EXISTING);
-			Runtime.getRuntime().load(path.toString());
-		} catch (NullPointerException | IOException e) {
-			Logger.info(e, "Error loading library from classpath: " + e);
-
-			// Try load from the Java system library path (-Djava.library.path)
-			System.loadLibrary(libName);
+		synchronized (LOADED_LIBRARIES) {
+			if (LOADED_LIBRARIES.get(libName) == null) {
+				try {
+					// First try load the library from within the JAR file
+					Path path = Files.createTempFile("lib" + libName, ".so");
+					path.toFile().deleteOnExit();
+					Files.copy(clz.getResourceAsStream("/lib/" + SystemInfo.getLibraryPath() + "/lib" + libName + ".so"), path,
+							StandardCopyOption.REPLACE_EXISTING);
+					Runtime.getRuntime().load(path.toString());
+				} catch (NullPointerException | IOException e) {
+					Logger.info(e, "Error loading library from classpath, trying System.loadLibrary: " + e);
+		
+					// Try load from the Java system library path (-Djava.library.path)
+					System.loadLibrary(libName);
+				}
+				LOADED_LIBRARIES.put(libName, LOADED);
+			}
 		}
 	}
 }
